sentiment-app/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
scikit-learn==1.5.2
pandas==2.2.3
nltk==3.9.1
joblib==1.4.2
numpy==2.1.2
matplotlib==3.9.2
wordcloud==1.9.3
tensorflow>=2.15.0
pillow==10.4.0
python-multipart==0.0.9

